refactor: constant -> config로 패키지 네이밍 변경
