+++ fourtoon-cookie/src/config/character.ts
