+++ fourtoon-cookie/src/pages/SignUpPage/AgreementInputLayout/AgreementInputLayout.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { View, Text, Pressable, Linking } from 'react-native';
 import * as S from "./AgreementInputLayout.style";
 import { MaterialIcons } from '@expo/vector-icons'; // 아이콘 사용
-import { APP_PRIVACY_URL, APP_TERMS_AGREEMENT_URL } from '../../../constants/appinfo';
+import { APP_PRIVACY_URL, APP_TERMS_AGREEMENT_URL } from '../../../config/appinfo';
 
 export interface AgreementInputLayoutProps {
     onAgreementChange: (isAgreed: boolean) => void;
